pyproject.toml
[project]
name = "mnist-project"
version = "0.1.0"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.7.2",
    "omegaconf>=2.3.0",
    "pytorch-lightning>=2.6.0",
    "scikit-learn>=1.8.0",
    "torch>=2.0",
    "torchvision>=0.15",
    "typer>=0.13.1",
    "wandb>=0.23.1",
]

[dependency-groups]
dev = [
    "coverage>=7.13.1",
    "pytest>=9.0.2",
]
